Add pack_scale_warn option to allow programs to throw away precision.
ExtentSeries.hpp: change int into size_t as it should have been.

ExtentType.?pp" add in new option and parse it.

Extent.cpp: just use format rather than boost::format (using has been present for a while)
  Switch to LintelLog::warn instead of cerr for warning about scaling, and check new
  option.

tests/CMakeLists.txt: add in new test.

pack-scale.cpp: Add test, verify we scale properly, also verify we only generate warnings
  when not disabled.

+++ include/DataSeries/ExtentSeries.hpp
@@ -306,7 +306,7 @@ public:
     void newRecord() {
 	INVARIANT(my_extent != NULL,
 		  "must set extent for data series before calling newRecord()");
-	int offset = my_extent->fixeddata.size();
+	size_t offset = my_extent->fixeddata.size();
 	my_extent->createRecords(1);
 	pos.cur_pos = my_extent->fixeddata.begin() + offset;
     }
@@ -320,7 +320,7 @@ public:
 	// leaves the current record position unchanged
 	INVARIANT(my_extent != NULL,
 		  "must set extent for data series before calling newRecord()\n");
-	int offset = pos.cur_pos - my_extent->fixeddata.begin();
+	size_t offset = pos.cur_pos - my_extent->fixeddata.begin();
 	my_extent->createRecords(nrecords);
 	pos.cur_pos = my_extent->fixeddata.begin() + offset;
     }
